refactor: add OCRClient
